pyproject.toml
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[project]
name = "devboard"
description = "A development dashboard for your projects."
authors = [{name = "Timothée Mazzucotelli", email = "dev@pawamoy.fr"}]
license = {text = "ISC"}
readme = "README.md"
requires-python = ">=3.8"
keywords = []
dynamic = ["version"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Documentation",
    "Topic :: Software Development",
    "Topic :: Utilities",
    "Typing :: Typed",
]
dependencies = [
    "textual>=0.40",
    "gitpython>=3.1",
    "appdirs>=1.4.4",
    "tomli>=2.0; python_version < '3.11'",
]

[project.urls]
Homepage = "https://pawamoy.github.io/devboard"
Documentation = "https://pawamoy.github.io/devboard"
Changelog = "https://pawamoy.github.io/devboard/changelog"
Repository = "https://github.com/pawamoy/devboard"
Issues = "https://github.com/pawamoy/devboard/issues"
Discussions = "https://github.com/pawamoy/devboard/discussions"
Gitter = "https://gitter.im/devboard/community"
Funding = "https://github.com/sponsors/pawamoy"

[project.scripts]
devboard = "devboard.cli:main"

[tool.pdm]
version = {source = "scm"}

[tool.pdm.build]
package-dir = "src"
editable-backend = "editables"
